+++ client/tws_foundation_client/lib/src/services/bases/trucks_service_base.dart
@@ -13,4 +13,7 @@ abstract class TrucksServiceBase extends CSMServiceBase {
 
   Effect<MigrationTransactionResult<Truck>> create(List<Truck> trucks, String auth);
 
+  Effect<MigrationUpdateResult<Truck>> update(Truck truck, String auth);
+
+
 }