+++ client/tws_foundation_client/test/integration/trucks_externals_service_test.dart
@@ -1,3 +1,5 @@
+import 'dart:math';
+
 import 'package:test/test.dart';
 import 'package:tws_foundation_client/tws_foundation_client.dart';
 
@@ -6,6 +8,32 @@ import '../integration_credentials.dart';
 void main() {
   late String auth;
   late TrucksExternalsServiceBase service;
+  late List<TruckExternal> mocks;
+
+  TruckExternal buildExternalTruck(String randomToken){
+    TruckCommon truckCommon = TruckCommon(
+      0, //id
+      1, //status
+      "ECO$randomToken", //economic
+      null, //location
+      null, //situation
+      null, //locationNavigation
+      null, //situationNavigation
+      null, //statusNavigation
+    );
+    TruckExternal truckExternal = TruckExternal(
+      0, //ID
+      1, //Status
+      0, //Common
+      null, //VIN
+      "Carrier $randomToken", 
+      "MEX$randomToken", 
+      "USA$randomToken", 
+      truckCommon, 
+      null //StatusNavigation
+    );
+    return truckExternal;
+  }
   setUp(
     () async {
       final TWSFoundationSource source = TWSFoundationSource(false);
@@ -27,6 +55,13 @@ void main() {
       );
 
       service = source.trucksExternals;
+      mocks = <TruckExternal>[];
+      for (int i = 0; i < 3; i++) {
+        int rnd = Random().nextInt(900)  + 99;
+        String randomToken = '${i}_qual$rnd';
+        TruckExternal mock = buildExternalTruck(randomToken);
+        mocks.add(mock);
+      }
     },
   );
 
@@ -56,6 +91,63 @@ void main() {
           expect(fact.page, 1);
           expect(fact.pages >= fact.page, true);
           expect(fact.sets.length, fact.records);
+        },
+      );
+    },
+  );
+  test(
+    'Create',
+    () async {
+      MainResolver<MigrationTransactionResult<TruckExternal>> fact = await service.create(mocks, auth);
+
+      bool resolved = false;
+      fact.resolve(
+        decoder: MigrationTransactionResultDecoder<TruckExternal>(TruckExternalDecoder()),
+        onException: (Object exception, StackTrace trace) => throw exception,
+        onConnectionFailure: () => throw Exception('Connection failure'),
+        onFailure: (FailureFrame failure, int status) => throw Exception(failure.estela.advise),
+        onSuccess: (SuccessFrame<MigrationTransactionResult<TruckExternal>> success) {
+          resolved = true;
+        },
+      );
+
+      expect(true, resolved, reason: 'The action wasn\'t resolved');
+    },
+  );
+
+  group(
+    'Update',
+    () {
+      final MigrationUpdateResultDecoder<TruckExternal> decoder = MigrationUpdateResultDecoder<TruckExternal>(TruckExternalDecoder());
+      late TruckExternal creationMock;
+      test(
+        'Creates when unexist',
+        () async {
+          int rnd = Random().nextInt(900)  + 99;
+          TruckExternal mock = buildExternalTruck("U_qual$rnd");
+
+          MainResolver<MigrationUpdateResult<TruckExternal>> fact = await service.update(mock, auth);
+          MigrationUpdateResult<TruckExternal> actEffect = await fact.act(decoder);
+          assert(actEffect.previous == null);
+          assert(actEffect.updated.id > 0);
+
+          creationMock = actEffect.updated;
+        },
+      );
+
+      test(
+        'Updates when exist',
+        () async {
+          int rnd = Random().nextInt(900)  + 99;
+          TruckExternal mock = creationMock.clone(vin: "UPDATEDVIN_T: $rnd");
+          mock.truckCommonNavigation!.economic = "UPDTECONOMIC:$rnd";
+          MainResolver<MigrationUpdateResult<TruckExternal>> fact = await service.update(mock, auth);
+          MigrationUpdateResult<TruckExternal> actEffect = await fact.act(decoder);
+          assert(actEffect.previous != null);
+          assert(actEffect.updated.id == creationMock.id);
+          assert(actEffect.updated.vin != null);
+          assert(actEffect.updated.truckCommonNavigation!.economic != actEffect.previous!.truckCommonNavigation!.economic);
+
         },
       );
     },